--- docker-compose-simple.yml
@@ -1,29 +1,0 @@
-# Simplest approach - command override
-services:
-  ollama:
-    image: ollama/ollama:latest
-    container_name: ollama
-    ports:
-      - "11434:11434"
-    volumes:
-      - ollama_data:/root/.ollama
-    environment:
-      - OLLAMA_HOST=0.0.0.0
-    deploy:
-      resources:
-        reservations:
-          devices:
-            - driver: nvidia
-              count: all
-              capabilities: [gpu]
-    restart: unless-stopped
-    command: >
-      sh -c "
-        ollama serve &
-        sleep 10 &&
-        ollama pull llama3.1:8b &&
-        ollama run llama3.1:8b
-      "
-
-volumes:
-  ollama_data:
